backend/requirements.txt
en_core_web_sm==3.7.1
feedparser==6.0.11
Requests==2.32.5
pymongo==4.10.1
python-dotenv==1.0.1
bcrypt==4.3.0
Flask==3.1.3
Flask-CORS==6.0.1
PyJWT==2.10.1
Pillow==10.4.0